https://mosipid.atlassian.net/browse/Merge bra

Former-commit-id: e9a529e1c949ffcba44acbac0a3354c9302450b4
Former-commit-id: 67e7bdc9a6280f669961bfb99a7e7de3b1ca21d2

+++ kernel/kernel-masterdata-service/src/main/java/io/mosip/kernel/masterdata/repository/LocationRepository.java
@@ -53,7 +53,7 @@ public interface LocationRepository extends BaseRepository<Location, CodeAndLang
 	 * 
 	 */
 	@Query(value = "FROM Location l where l.langCode=?1 and l.hierarchyLevel >=?2 and (l.isDeleted is null or l.isDeleted=false)")
-	List<Location> getAllLocationsByLangCodeAndLevel(String langCode, Integer level);
+	List<Location> getAllLocationsByLangCodeAndLevel(String langCode, Short level);
 
 	/**
 	 * checks whether the location name is valid location or not
